requirements.txt
streamlit==1.35.0
mysql-connector-python==8.4.0
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
numpy==1.26.4
streamlit_option_menu==0.3.13
plotly==5.22.0
